Changes: 1. Add validation. 2. Bug fixes 3. Part of Javadoc

+++ src/main/java/com/example/sport_store/controller/admin/AdminController.java
@@ -3,83 +3,107 @@ package com.example.sport_store.controller.admin;
 import com.example.sport_store.entity.*;
 import com.example.sport_store.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.validation.Valid;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.security.Principal;
 import java.util.Date;
-import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping("/admin")
 public class AdminController {
     private static final String UPLOAD_FOLDER = "upload/customer_image/";
-    @Autowired
-    RoleRepository roleRepository;
-    @Autowired
-    CustomerRepository customerRepository;
-    @Autowired
-    AssignedRoleRepository assignedRoleRepository;
-    @Autowired
-    OrderStatusRepository orderStatusRepository;
-    @Autowired
-    PaymentTypeRepository paymentTypeRepository;
+    private static final Integer MAX_ROWS_PER_PAGE = 5;
+    final RoleRepository roleRepository;
+    final CustomerRepository customerRepository;
+    final AssignedRoleRepository assignedRoleRepository;
+    final OrderStatusRepository orderStatusRepository;
+    final PaymentTypeRepository paymentTypeRepository;
+    final OrderRepository orderRepository;
 
+    @Autowired
+    public AdminController(RoleRepository roleRepository,
+                           CustomerRepository customerRepository,
+                           AssignedRoleRepository assignedRoleRepository,
+                           OrderStatusRepository orderStatusRepository,
+                           PaymentTypeRepository paymentTypeRepository,
+                           OrderRepository orderRepository) {
+        this.roleRepository = roleRepository;
+        this.customerRepository = customerRepository;
+        this.assignedRoleRepository = assignedRoleRepository;
+        this.orderStatusRepository = orderStatusRepository;
+        this.paymentTypeRepository = paymentTypeRepository;
+        this.orderRepository = orderRepository;
+    }
+
+    /**
+     * Определение объектов, которые должны быть частью модели(Model).
+     */
     @ModelAttribute
     public void addAttributes(Model model, Principal principal) {
         model.addAttribute("paymentTypes", paymentTypeRepository.findAll());
         Customer customer = customerRepository.findUserByLogin(principal.getName());
-        model.addAttribute("user",
-                customer.getSurname() + " " + customer.getName());
+        model.addAttribute("user", customer.getSurname() + " " + customer.getName());
     }
 
+    /**
+     * Обработка запроса - главная страница интерфейса администратора.
+     */
     @GetMapping
-    public String adminPage() {
+    public String adminMainPageView() {
         return "admin/admin";
     }
 
+    /**
+     * Обработка запроса - вывод списка всех сотрудников.
+     */
     @GetMapping("/employees")
-    public String employeesView(Model model, @RequestParam(defaultValue = "0") int page) {
-        List<AssignedRole> assignedRoleList = assignedRoleRepository
-                .findAll()
-                .stream()
-                .filter(assignedRole -> assignedRole.getRole() != roleRepository.getRoleByName("ROLE_USER"))
-                .collect(Collectors.toList());
-        Page<AssignedRole> assignedRolePage =
-                new PageImpl<>(assignedRoleList, PageRequest.of(page, 10), assignedRoleList.size());
-        model.addAttribute("employees", assignedRolePage);
+    public String employeesPageView(Model model, @RequestParam(defaultValue = "0") int page, Principal principal) {
+        model.addAttribute("employees",
+                assignedRoleRepository
+                        .findAssignedRolesWithoutRoleUserAndPrincipal("ROLE_USER",
+                                principal.getName(), PageRequest.of(page, MAX_ROWS_PER_PAGE)));
         model.addAttribute("currentPage", page);
         return "admin/employees";
     }
 
+    /**
+     * Обработка запроса - вывод списка всех покупателей.
+     */
     @GetMapping("/customers")
     public String customersView(Model model, @RequestParam(defaultValue = "0") int page) {
         model.addAttribute("customers",
                 assignedRoleRepository.findAssignedRolesByRole(roleRepository.getRoleByName("ROLE_USER"),
-                        PageRequest.of(page, 10)));
+                        PageRequest.of(page, MAX_ROWS_PER_PAGE)));
         model.addAttribute("currentPage", page);
         return "admin/customers";
     }
 
+    /**
+     * Обработка запроса - создать нового сотрудника.
+     */
     @GetMapping("/createEmployee")
     public String createEmployee(Model model) {
-        model.addAttribute("employee", new Customer());
+        model.addAttribute("customer", new Customer());
         model.addAttribute("roles", roleRepository.findAll());
         return "admin/employee";
     }
 
+    /**
+     * Обработка запроса - создать нового покупателя.
+     */
     @GetMapping("/createCustomer")
     public String createCustomer(Model model) {
         model.addAttribute("customer", new Customer());
@@ -87,10 +111,34 @@ public class AdminController {
     }
 
     @PostMapping("/saveEmployee")
-    public String saveEmployee(Customer customer,
+    public String saveEmployee(@Valid Customer customer,
+                               BindingResult result,
+                               Model model,
                                MultipartFile image,
                                @RequestParam String roleId,
-                               @RequestParam String password) {
+                               @RequestParam String password,
+                               @RequestParam String email,
+                               @RequestParam String phone,
+                               @RequestParam String login) {
+        if (customerRepository.findCustomerByLogin(login).orElse(null) != null) {
+            FieldError error = new FieldError("customer", "login",
+                    "Пользователь с таким логином уже существует");
+            result.addError(error);
+        }
+        if (customerRepository.findCustomerByEmail(email).orElse(null) != null) {
+            FieldError error = new FieldError("customer", "email",
+                    "Пользователь с таким почтовым адресом уже существует");
+            result.addError(error);
+        }
+        if (customerRepository.findCustomerByPhone(phone).orElse(null) != null) {
+            FieldError error = new FieldError("customer", "phone",
+                    "Пользователь с таким номером телефона уже существует");
+            result.addError(error);
+        }
+        if (result.hasErrors()) {
+            model.addAttribute("roles", roleRepository.findAll());
+            return "admin/employee";
+        }
         if (!image.isEmpty()) {
             try {
                 String newImageFileName = UUID.randomUUID().toString().replaceAll("-", "") +
@@ -111,9 +159,31 @@ public class AdminController {
     }
 
     @PostMapping("/saveCustomer")
-    public String saveCustomer(Customer customer,
+    public String saveCustomer(@Valid Customer customer,
+                               BindingResult result,
                                MultipartFile image,
-                               @RequestParam String password) {
+                               @RequestParam String password,
+                               @RequestParam String email,
+                               @RequestParam String phone,
+                               @RequestParam String login) {
+        if (customerRepository.findCustomerByLogin(login).orElse(null) != null) {
+            FieldError error = new FieldError("customer", "login",
+                    "Пользователь с таким логином уже существует");
+            result.addError(error);
+        }
+        if (customerRepository.findCustomerByEmail(email).orElse(null) != null) {
+            FieldError error = new FieldError("customer", "email",
+                    "Пользователь с таким почтовым адресом уже существует");
+            result.addError(error);
+        }
+        if (customerRepository.findCustomerByPhone(phone).orElse(null) != null) {
+            FieldError error = new FieldError("customer", "phone",
+                    "Пользователь с таким номером телефона уже существует");
+            result.addError(error);
+        }
+        if (result.hasErrors()) {
+            return "admin/customer";
+        }
         if (!image.isEmpty()) {
             try {
                 String newImageFileName = UUID.randomUUID().toString().replaceAll("-", "") +
@@ -136,7 +206,7 @@ public class AdminController {
     @GetMapping("/editEmployee")
     public String editEmployee(Long id, Model model) {
         Customer customer = customerRepository.getCustomerById(id);
-        model.addAttribute("employee", customer);
+        model.addAttribute("customer", customer);
         model.addAttribute("roles", roleRepository.findAll());
         model.addAttribute("employeeRole",
                 assignedRoleRepository.findAssignedRolesByCustomer(customer).get(0).getRole());
@@ -150,7 +220,9 @@ public class AdminController {
     }
 
     @PostMapping("/updateEmployee")
-    public String updateEmployee(MultipartFile image,
+    public String updateEmployee(@Valid Customer customer,
+                                 MultipartFile image,
+                                 BindingResult result,
                                  @RequestParam Long id,
                                  @RequestParam String name,
                                  @RequestParam String surname,
@@ -160,7 +232,34 @@ public class AdminController {
                                  @RequestParam String city,
                                  @RequestParam String login,
                                  @RequestParam String roleId) {
-        Customer customer = customerRepository.getCustomerById(id);
+        Customer checkCustomer = customerRepository.findCustomerByLogin(login).orElse(null);
+        if (checkCustomer != null) {
+            if (!checkCustomer.getId().equals(id)) {
+                FieldError error = new FieldError("customer", "login",
+                        "Пользователь с таким логином уже существует");
+                result.addError(error);
+            }
+        }
+        checkCustomer = customerRepository.findCustomerByEmail(email).orElse(null);
+        if (checkCustomer != null) {
+            if (!checkCustomer.getId().equals(id)) {
+                FieldError error = new FieldError("customer", "email",
+                        "Пользователь с таким почтовым адресом уже существует");
+                result.addError(error);
+            }
+        }
+        checkCustomer = customerRepository.findCustomerByPhone(phone).orElse(null);
+        if (checkCustomer != null) {
+            if (!checkCustomer.getId().equals(id)) {
+                FieldError error = new FieldError("customer", "phone",
+                        "Пользователь с таким номером телефона уже существует");
+                result.addError(error);
+            }
+        }
+        if (result.hasErrors()) {
+            return "admin/edit_employee";
+        }
+        customer = customerRepository.getCustomerById(id);
         if (!image.isEmpty()) {
             try {
                 String newImageFileName = UUID.randomUUID().toString().replaceAll("-", "") +
@@ -186,8 +285,10 @@ public class AdminController {
     }
 
     @PostMapping("/updateCustomer")
-    public String updateCustomer(MultipartFile image,
-                                 @RequestParam Long id,
+    public String updateCustomer(@Valid Customer customer,
+                                 MultipartFile image,
+                                 BindingResult result,
+                                 @RequestParam String id,
                                  @RequestParam String name,
                                  @RequestParam String surname,
                                  @RequestParam String phone,
@@ -195,7 +296,34 @@ public class AdminController {
                                  @RequestParam String country,
                                  @RequestParam String city,
                                  @RequestParam String login) {
-        Customer customer = customerRepository.getCustomerById(id);
+        Customer checkCustomer = customerRepository.findCustomerByLogin(login).orElse(null);
+        if (checkCustomer != null) {
+            if (!checkCustomer.getId().equals(Long.parseLong(id))) {
+                FieldError error = new FieldError("customer", "login",
+                        "Пользователь с таким логином уже существует");
+                result.addError(error);
+            }
+        }
+        checkCustomer = customerRepository.findCustomerByEmail(email).orElse(null);
+        if (checkCustomer != null) {
+            if (!checkCustomer.getId().equals(Long.parseLong(id))) {
+                FieldError error = new FieldError("customer", "email",
+                        "Пользователь с таким почтовым адресом уже существует");
+                result.addError(error);
+            }
+        }
+        checkCustomer = customerRepository.findCustomerByPhone(phone).orElse(null);
+        if (checkCustomer != null) {
+            if (!checkCustomer.getId().equals(Long.parseLong(id))) {
+                FieldError error = new FieldError("customer", "phone",
+                        "Пользователь с таким номером телефона уже существует");
+                result.addError(error);
+            }
+        }
+        if (result.hasErrors()) {
+            return "admin/edit_customer";
+        }
+        customer = customerRepository.getCustomerById(Long.parseLong(id));
         if (!image.isEmpty()) {
             try {
                 String newImageFileName = UUID.randomUUID().toString().replaceAll("-", "") +
@@ -248,16 +376,21 @@ public class AdminController {
 
     @PostMapping("/saveOrderStatus")
     public String saveOrderStatus(@RequestParam String id, @RequestParam String name) {
-        OrderStatus orderStatus = id.isEmpty() ? new OrderStatus() :
-                orderStatusRepository.getOrderStatusById(Long.parseLong(id));
-        orderStatus.setName(name);
-        orderStatusRepository.save(orderStatus);
+        if (orderStatusRepository.findOrderStatusByName(name).orElse(null) == null) {
+            OrderStatus orderStatus = id.isEmpty() ? new OrderStatus() :
+                    orderStatusRepository.getOrderStatusById(Long.parseLong(id));
+            orderStatus.setName(name);
+            orderStatusRepository.save(orderStatus);
+        }
         return "redirect:/admin/orderStatuses";
     }
 
     @GetMapping("/deleteOrderStatus")
     public String deleteOrderStatus(Long id) {
-        orderStatusRepository.deleteById(id);
+        if (orderRepository
+                .findOrderByOrderStatus(orderStatusRepository.findById(id).orElse(null)).isEmpty()) {
+            orderStatusRepository.deleteById(id);
+        }
         return "redirect:/admin/orderStatuses";
     }
 
@@ -276,16 +409,21 @@ public class AdminController {
 
     @PostMapping("/savePaymentType")
     public String savePaymentType(@RequestParam String id, @RequestParam String name) {
-        PaymentType paymentType = id.isEmpty() ? new PaymentType() :
-                paymentTypeRepository.getPaymentTypeById(Long.parseLong(id));
-        paymentType.setName(name);
-        paymentTypeRepository.save(paymentType);
+        if (paymentTypeRepository.findPaymentTypeByName(name).orElse(null) == null) {
+            PaymentType paymentType = id.isEmpty() ? new PaymentType() :
+                    paymentTypeRepository.getPaymentTypeById(Long.parseLong(id));
+            paymentType.setName(name);
+            paymentTypeRepository.save(paymentType);
+        }
         return "redirect:/admin/paymentTypes";
     }
 
     @GetMapping("/deletePaymentType")
     public String deletePaymentType(Long id) {
-        paymentTypeRepository.deleteById(id);
+        if (orderRepository
+                .findOrderByPaymentType(paymentTypeRepository.findById(id).orElse(null)).isEmpty()) {
+            paymentTypeRepository.deleteById(id);
+        }
         return "redirect:/admin/paymentTypes";
     }
